CI: Add github actions workflow automations
Adds a `compile.yml` GitHub Actions workflow which will automatically
compile all collections in the repository into compendiums, zip them up
into a single archive and then upload the archive as an artifact.

If a tag was pushed, a release will be automatically created for that
tag with all of the uncompressed compiled compendium files attached as
assets to the release.

The above workflow will run automatically on any push or pull request.
It can additionally be manually invoked through the "workflow_dispatch"
trigger.