REFACTOR: Remove SCAG duplicate spells
All of the spells which are in the Sword Coast Adventurer's Guide are
duplicated elsewhere (such as in Tasha's Cauldron of Everything), so
they are no longer needed to be included, and including them causes
duplicate spells to show up. Therefore they have been removed.

BREAKING CHANGE: Any workflow which expects spells-scag.xml will break.